Simplified main menu and improved background
Main menu background now adds an offset to the menu background box instead of manually moving the vertices
Main menu no longer generates circles, and a separate class for the background sprite is no longer required. Instead, two shaders are used. One for the background box inside the stencil, and a separate one for the general background.

Finally, the sliding has been adjusted on several buttons so it doesn't cover the entire background, which I think looks nicer