Removed naming convention of grunt/soldier etc
This worked for the design of the game, but made it hard to differentiate between enemies. As the design evolved to include more than 1 type of grunt, soldier etc, it became harder to differentiate in code too. This also resulted in ugly naming conventions of "gruntCharger", "soldierDrone". Naming them just "charger" and "drone" makes them visually distinct in code from the name alone, without having to look back at the classes, and the class naming convention can be kept in a potential bestiary.