Fixed eyes and tails appearing at the center of the map
When an enemy spawns, there is a very brief period (1 frame maybe) where the eye and tail positions are set to 0, 0. Parameters have been added to the constructor so that when an enemy spawns, it can get the starting point for the eye and tail position vectors to where the enemy spawns. Additional effects can be added later to cover this up