Refactored enemy spawning logic
The enemy spawn warning is now a dedicated class which handles the spawning of a specific enemy, removing the cluttered rendering and spawning logic from the stage director.

For now, the stage director acts on a timer. After it spawns a group of enemy warnings, it sets this timer to 3 (the time it takes for the warnings to spawn an enemy). After this time, enemy kills will be registered with the stage director. If the number of kills after this time becomes greater than the minimum kills for the next wave, the next wave will start.