Added "next wave conditions" to the stage director
Next wave conditions are conditions that must be met to start the next wave. Two are currently defined - minimum kills, and a timer. If at least a custom timer isn't specified, a default override of a 15 seconds timer is used. For example, minimum kills can be enabled, but if these kills can't be met, the next wave won't start, so a timer is always needed in some capacity.